feat: moved secondary functions to separate modules
